Add information about how to translate the docs to the contributor's manual.
Summary:
This is a bit verbose in everything that happens to the translations, but I'd
like future translators to be able to visualize how they would go about joining
a team and then seeing their translations get added to the manual.

Test Plan: Read the entry and see if there's any errors in it.

Reviewers: aacid, ltoscano, #krita_manual, rempt

Reviewed By: #krita_manual, rempt

Subscribers: scottpetrovic, rempt

Tags: #krita_manual

Differential Revision: https://phabricator.kde.org/D19227