Port the modern cpp and intro to hacking krita from the community wiki to contributor's manual.
Summary:
In an attempt to do some moving in the creation of the contributor's manual, I've
decided to port these two pages.

https://community.kde.org/Krita/C%2B%2B11
https://community.kde.org/Krita/Hacking

When these are merged we can clear the pages and point at the contributor manual versions.

Test Plan: Check if everything's still relevant? Did I make a mistake?

Reviewers: rempt, dkazakov, jounip, tymond, #krita_manual

Tags: #krita_manual

Differential Revision: https://phabricator.kde.org/D20250