Remove mentions of draft, disable WebIDE picture
Now all changes should go to master, so no need to mention draft at all.
WebIDE picture was wrong, not only did it mention draft, but also it shows
the official repository as a correct repository to be in when you're using this mode,
which is incorrect (you should be using your fork).