Mention ZSync file on the "Making a release" page
Krita 5.0 includes the built-in update mechanism for AppImages. The
release CI job will provide one more file: the zsync metadata file
needed for the update mechanism to work.

The zsync metadata file is mentioned in the list of files to upload to
download.kde.org in step 12 and the total number of files is updated.