Try to split search term into smaller parts for zh and ja
When sphinx generates the search index, terms gets split into the
smallest logical part, for example "笔刷预设介绍" will be split into
three individual terms - "笔刷", "预设" and "介绍". The search page
JavaScript does not know how to do segmentation (wouldn't be feasible
anyway due to the need of a dictionary). Therefore here we add an
extra logic to attempt to further split the search terms according to
available terms in the search index to make the search function more
useful for Chinese and Japanese languages.

This logic requires that every part of the search term to be an existing
term in the index, like "笔刷预设介绍". If the search term is instead
"笔刷预设道路" and that "道路" does not exist in the index, this logic
will not apply the split and the search will yield no results.

BUG: 440246