Add "right click to undo last point" info to polygon-like tools
This adds info to polygon, polygonal selection, path selection and freehand selection tools to indicate that clicking with the right mouse button will undo the last generated point in the path.

Also, in the freehand selection tool, the use of <kbd>ctrl</kbd> was marked as deprecated and te text said that the feature was removed in 5.1. But it wasn't, so I removed both that text and the deprecated directive.